css/image-fit/js/inherited/005.html

<!doctype html>
<title>'inherit'ing from &lt;svg></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var e = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
  e.appendChild(document.createElement('div'));
  e.firstChild.style.OObjectFit = 'inherit';
  e.firstChild.style.OObjectPosition = 'inherit';
  document.body.appendChild(e);
  assert_equals(getComputedStyle(e.firstChild, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e.firstChild, '').OObjectPosition, '50% 50%');
});
</script>
